+++ db/migrate/20260526150000_add_prusalink_reachable_to_printers.rb
@@ -1,0 +1,8 @@
+class AddPrusalinkReachableToPrinters < ActiveRecord::Migration[8.1]
+  def change
+    change_table :printers, bulk: true do |t|
+      t.boolean :prusalink_reachable
+      t.datetime :prusalink_checked_at
+    end
+  end
+end