+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_05_25_160100) do
+ActiveRecord::Schema[8.1].define(version: 2026_05_26_150000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -87,7 +87,9 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_25_160100) do
     t.string "model"
     t.string "name", null: false
     t.string "operational_state", default: "unknown", null: false
+    t.datetime "prusalink_checked_at"
     t.text "prusalink_key"
+    t.boolean "prusalink_reachable"
     t.datetime "updated_at", null: false
     t.index ["name"], name: "index_printers_on_name", unique: true
   end